+++ src/IRFeedbackLED.hpp
@@ -36,6 +36,8 @@
  * @{
  */
 
+#include "digitalWriteFast.h"
+
 /**
  * Contains pin number and enable status of the feedback LED
  */
@@ -157,4 +159,3 @@ void setBlinkPin(uint8_t aBlinkPin) {
 /** @}*/
 
 #endif // _IR_FEEDBACK_LED_HPP
-#pragma once