+++ src/IRProtocol.h
@@ -101,4 +101,3 @@ const __FlashStringHelper* getProtocolString(decode_type_t aProtocol);
 #define LEGO_MODE_SINGLE    0x4 // here the 2 LSB have meanings like Output A / Output B
 
 #endif // _IR_PROTOCOL_H
-#pragma once