+++ src/IRSend.hpp
@@ -71,7 +71,7 @@ void IRsend::begin(){
 /**
  * @param aFeedbackLEDPin if 0, then take board specific FEEDBACK_LED_ON() and FEEDBACK_LED_OFF() functions
  */
-void IRsend::begin(bool aEnableLEDFeedback, uint8_t aFeedbackLEDPin) {
+void IRsend::begin(bool aEnableLEDFeedback, uint_fast8_t aFeedbackLEDPin) {
 #if !defined(NO_LED_FEEDBACK_CODE)
     bool tEnableLEDFeedback = DO_NOT_ENABLE_LED_FEEDBACK;
     if(aEnableLEDFeedback) {
@@ -85,7 +85,7 @@ void IRsend::begin(bool aEnableLEDFeedback, uint8_t aFeedbackLEDPin) {
 }
 
 #else // defined(IR_SEND_PIN)
-IRsend::IRsend(uint8_t aSendPin) { // @suppress("Class members should be properly initialized")
+IRsend::IRsend(uint_fast8_t aSendPin) { // @suppress("Class members should be properly initialized")
     sendPin = aSendPin;
 #  if !defined(NO_LED_FEEDBACK_CODE)
     setLEDFeedback(0, DO_NOT_ENABLE_LED_FEEDBACK);
@@ -96,14 +96,14 @@ IRsend::IRsend(uint8_t aSendPin) { // @suppress("Class members should be properl
  * Initializes the send pin and enable LED feedback with board specific FEEDBACK_LED_ON() and FEEDBACK_LED_OFF() functions
  * @param aSendPin The Arduino pin number, where a IR sender diode is connected.
  */
-void IRsend::begin(uint8_t aSendPin) {
+void IRsend::begin(uint_fast8_t aSendPin) {
     sendPin = aSendPin;
 #  if !defined(NO_LED_FEEDBACK_CODE)
     setLEDFeedback(USE_DEFAULT_FEEDBACK_LED_PIN, LED_FEEDBACK_ENABLED_FOR_SEND);
 #  endif
 }
 
-void IRsend::setSendPin(uint8_t aSendPin) {
+void IRsend::setSendPin(uint_fast8_t aSendPin) {
     sendPin = aSendPin;
 }
 #endif // defined(IR_SEND_PIN)
@@ -113,7 +113,7 @@ void IRsend::setSendPin(uint8_t aSendPin) {
  * @param aSendPin The Arduino pin number, where a IR sender diode is connected.
  * @param aFeedbackLEDPin if 0, then take board specific FEEDBACK_LED_ON() and FEEDBACK_LED_OFF() functions
  */
-void IRsend::begin(uint8_t aSendPin, bool aEnableLEDFeedback, uint8_t aFeedbackLEDPin) {
+void IRsend::begin(uint_fast8_t aSendPin, bool aEnableLEDFeedback, uint_fast8_t aFeedbackLEDPin) {
 #if defined(IR_SEND_PIN)
     (void) aSendPin; // for backwards compatibility
 #else
@@ -293,7 +293,7 @@ void IRsend::sendRaw_P(const uint16_t aBufferWithMicroseconds[], uint_fast8_t aL
      * Raw data starts with a mark
      */
     for (uint_fast8_t i = 0; i < aLengthOfBuffer; i++) {
-        uint16_t duration = pgm_read_word(&aBufferWithMicroseconds[i]);
+        unsigned int  duration = pgm_read_word(&aBufferWithMicroseconds[i]);
         if (i & 1) {
             // Odd
             space(duration);
@@ -317,7 +317,7 @@ void IRsend::sendRaw_P(const uint8_t aBufferWithTicks[], uint_fast8_t aLengthOfB
     enableIROut(aIRFrequencyKilohertz);
 
     for (uint_fast8_t i = 0; i < aLengthOfBuffer; i++) {
-        uint16_t duration = pgm_read_byte(&aBufferWithTicks[i]) * (uint16_t) MICROS_PER_TICK;
+        unsigned int  duration = pgm_read_byte(&aBufferWithTicks[i]) * (unsigned int ) MICROS_PER_TICK;
         if (i & 1) {
             // Odd
             space(duration);
@@ -334,7 +334,7 @@ void IRsend::sendRaw_P(const uint8_t aBufferWithTicks[], uint_fast8_t aLengthOfB
  * The output always ends with a space
  */
 void IRsend::sendPulseDistanceWidthData(unsigned int aOneMarkMicros, unsigned int aOneSpaceMicros, unsigned int aZeroMarkMicros,
-        unsigned int aZeroSpaceMicros, uint32_t aData, uint8_t aNumberOfBits, bool aMSBfirst, bool aSendStopBit) {
+        unsigned int aZeroSpaceMicros, uint32_t aData, uint_fast8_t aNumberOfBits, bool aMSBfirst, bool aSendStopBit) {
 
     if (aMSBfirst) {  // Send the MSB first.
         // send data from MSB to LSB until mask bit is shifted out
@@ -381,7 +381,7 @@ void IRsend::sendBiphaseData(unsigned int aBiphaseTimeUnit, uint32_t aData, uint
     mark(aBiphaseTimeUnit);
 
     IR_TRACE_PRINT('S');
-    uint8_t tLastBitValue = 1; // Start bit is a 1
+    uint_fast8_t tLastBitValue = 1; // Start bit is a 1
 
 // Data - Biphase code MSB first
     for (uint32_t tMask = 1UL << (aNumberOfBits - 1); tMask; tMask >>= 1) {
@@ -606,7 +606,7 @@ void IRsend::customDelayMicroseconds(unsigned long aMicroseconds) {
  * A few hours staring at the ATmega documentation and this will all make sense.
  * See my Secrets of Arduino PWM at http://www.righto.com/2009/07/secrets-of-arduino-pwm.html for details.
  */
-void IRsend::enableIROut(uint8_t aFrequencyKHz) {
+void IRsend::enableIROut(uint_fast8_t aFrequencyKHz) {
 #if defined(SEND_PWM_BY_TIMER)
     timerConfigForSend(aFrequencyKHz); // must set output pin mode and disable receive interrupt if required, e.g. uses the same resource
 
@@ -637,4 +637,3 @@ unsigned int IRsend::getPulseCorrectionNanos() {
 
 /** @}*/
 #endif // _IR_SEND_HPP
-#pragma once