+++ src/private/IRTimer.hpp
@@ -1599,4 +1599,3 @@ void timerConfigForSend(uint8_t aFrequencyKHz) {
 /** @}*/
 /** @}*/
 #endif // _IR_TIMER_HPP
-#pragma once